library-flask/requirements.txt
click==8.1.3
Flask==3.0.3
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.3.0
protobuf==4.21.1
python-dotenv==0.19.2
Werkzeug==3.0.2
zipp==3.7.0
